refactor: new middleware thing
