+++ package.json
@@ -28,7 +28,7 @@
     "@happykit/flags": "^2.0.5-perf.0",
     "@sendgrid/mail": "7.6.2",
     "@sentry/cli": "1.74.4",
-    "@sentry/nextjs": "6.19.7",
+    "@sentry/nextjs": "7.5.1",
     "@stripe/react-stripe-js": "1.7.2",
     "@stripe/stripe-js": "1.29.0",
     "@stripe/terminal-js": "0.10.0",
@@ -40,7 +40,7 @@
     "framer-motion": "6.3.3",
     "micro": "9.3.4",
     "micro-cors": "0.1.1",
-    "next": "12.1.6",
+    "next": "12.2.0",
     "next-pwa": "5.5.4",
     "next-seo": "5.4.0",
     "qrcode": "1.5.0",
@@ -58,7 +58,7 @@
   },
   "devDependencies": {
     "@fast-csv/parse": "4.3.6",
-    "@next/eslint-plugin-next": "12.1.6",
+    "@next/eslint-plugin-next": "12.2.0",
     "@types/micro": "7.3.7",
     "@types/micro-cors": "0.1.2",
     "@types/node": "17.0.18",
